feat: Move setCallbackFns
