+++ test/shared/utils.spec.js
@@ -1,4 +1,4 @@
-import { deref, hasOwn, isAbsoluteUri, validateCallbackFn } from '@/shared/utils'
+import { deref, hasOwn, isAbsoluteUri, setCallbackFns, validateCallbackFn } from '@/shared/utils'
 
 test('dereferences objects', () => {
   const testObj = { foo: 'bar' }
@@ -49,4 +49,32 @@ test('is falsy on invalid callback', () => {
   expect(validateCallbackFn(true)).toBeFalsy()
   expect(validateCallbackFn(42)).toBeFalsy()
   expect(validateCallbackFn('str')).toBeFalsy()
+})
+
+test('returns valid arrays', () => {
+  const cbArr1 = [
+    () => {}
+  ]
+
+  const cbArr2 = [
+    (foo) => { return foo },
+    (bar) => { return bar }
+  ]
+
+  expect(setCallbackFns(cbArr1)).toStrictEqual(cbArr1)
+  expect(setCallbackFns(cbArr2)).toStrictEqual(cbArr2)
+})
+
+test('throws on invalid arrays', () => {
+  const cbArr3 = [
+    (foo) => { return foo },
+    42,
+    (bar) => { return bar }
+  ]
+
+  function testInvalid () {
+    setCallbackFns(cbArr3)
+  }
+
+  expect(testInvalid).toThrowErrorMatchingSnapshot()
 })
